Create Working with formatting
This was the first problem that showed us how to use the format command. The command was used in this situation, building off of the previous assignment, but be able to format decimal places. Also teaching us how to use things like formatting commas in proper positions for large numbers and decimal places.